refactor: change money floats to int64 (QA-P1-01)

+++ web/views/assets_templ.go
@@ -13,7 +13,7 @@ import (
 	"github.com/cecep-azhar/jurnalumi/internal/models"
 )
 
-func AssetManagement(tenant models.Tenant, user models.User, assets []models.CommodityAsset, totalAssetValue float64) templ.Component {
+func AssetManagement(tenant models.Tenant, user models.User, assets []models.CommodityAsset, totalAssetValue int64) templ.Component {
 	return templruntime.GeneratedTemplate(func(templ_7745c5c3_Input templruntime.GeneratedComponentInput) (templ_7745c5c3_Err error) {
 		templ_7745c5c3_W, ctx := templ_7745c5c3_Input.Writer, templ_7745c5c3_Input.Context
 		if templ_7745c5c3_CtxErr := ctx.Err(); templ_7745c5c3_CtxErr != nil {
@@ -42,7 +42,7 @@ func AssetManagement(tenant models.Tenant, user models.User, assets []models.Com
 	})
 }
 
-func assetContent(assets []models.CommodityAsset, totalAssetValue float64) templ.Component {
+func assetContent(assets []models.CommodityAsset, totalAssetValue int64) templ.Component {
 	return templruntime.GeneratedTemplate(func(templ_7745c5c3_Input templruntime.GeneratedComponentInput) (templ_7745c5c3_Err error) {
 		templ_7745c5c3_W, ctx := templ_7745c5c3_Input.Writer, templ_7745c5c3_Input.Context
 		if templ_7745c5c3_CtxErr := ctx.Err(); templ_7745c5c3_CtxErr != nil {
